+++ controller/cartController.js
@@ -1,15 +1,34 @@
-const async = require("../units/error")
-const cartService = require("../services/cartsService")
+const { catchAsync } = require("../utils/error");
+const cartService = require("../services/cartService");
 
-const selectData = catchAsync(async(req,res) =>{
-    const {
-        userId,
-        plantId,
-        plantQuantity,
-        potId,
-        potQuantity,
-        nutrientId,
-        nutrientQuantity,
-      } = req.body;
-    
-})
+const insertData = catchAsync(async (req, res) => {
+  const {
+    userId,
+    plantId,
+    plantQuantity,
+    potId,
+    potQuantity,
+    nutrientId,
+    nutrientQuantity,
+  } = req.body;
+
+  if (!userId) {
+    throw new Error("KEY_ERROR");
+  }
+
+  await cartService.insertData(
+    userId,
+    plantId,
+    plantQuantity,
+    potId,
+    potQuantity,
+    nutrientId,
+    nutrientQuantity
+  );
+
+  res.status(201).json({ message: "upsert data to cart success!!" });
+});
+
+module.exports = {
+  insertData,
+};
