+++ routes/cartsRouter.js
@@ -1,2 +1,10 @@
 const express = require("express");
-const cartController = require("../")
+const cartController = require("../controllers/cartController");
+
+const router = express.Router();
+
+router.post("", cartController.insertData);
+
+module.exports = {
+  router,
+};
